envolve components with user info
