+++ src/App.jsx
@@ -11,8 +11,64 @@ import Login from "./components/login/Login";
 import MyProfile from "./components/myProfile/myProfile";
 import PaymentGateway from "./components/paymentGateway/PaymentGateway";
 
+import React, { useEffect, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
+
+import { createUser, typeUser, getCountry } from "./redux/action/actions";
+import { useAuth0 } from "@auth0/auth0-react";
+
 
 const App = () => {
+  const dispatch = useDispatch();
+
+  const userAuth = useSelector((state) => state.user)
+
+  const [token, setToken] = useState()
+  const { isAuthenticated, user, getIdTokenClaims, logout, loginWithRedirect } = useAuth0()
+
+  useEffect(() => {
+    const fetchUserInformation = async () => {
+      try {
+        const idTokenClaims = await getIdTokenClaims();
+        const idToken = idTokenClaims?.__raw;
+        setToken(idToken);
+      } catch (error) {
+        console.error('Error fetching id token:', error);
+      }
+    };
+
+    if (isAuthenticated) {
+      fetchUserInformation();
+    }
+  }, [isAuthenticated, getIdTokenClaims]);
+
+  useEffect(() => {
+    const fetchUser = async () => {
+      if (token) {
+        console.log(user?.email);
+        await dispatch(createUser(user?.email, token));
+      }
+    };
+
+    if (token) {
+      fetchUser();
+    }
+  }, [token, user]);
+
+  useEffect(() => {
+    if (userAuth?.email) {
+      console.log(userAuth.typeUser);
+      dispatch(typeUser(userAuth.typeUser));
+
+    }
+  }, [userAuth]);
+
+  useEffect(() => {
+    if (userAuth?.CountryId) {
+      console.log(userAuth?.CountryId);
+      dispatch(getCountry(userAuth?.CountryId));
+    }
+  }, [userAuth]);
 
   return (
     <div>
